u-root: -defaultsh for default shell to start after init.
Fixes #748.

E.g.:

  ./u-root -defaultsh=elvish github.com/u-root/elvish \
                      github.com/u-root/u-root/cmds/*

  ./u-root -defaultsh=/bin/bash -files=/bin/bash

This will create a symlink /bin/defaultsh pointing to the appropriate
shell.

In source mode, init will resolve one level of the symlink before
executing the command in order to achieve the correct argv[0]. E.g. if
the symlink chain is:

  /buildbin/defaultsh -> /buildbin/rush -> /buildbin/installcommand

Then to compile rush, argv[0] has to have the suffix "/rush".