initramfs: add directories' children directly.
Previously, directories' children were added when the archive was
written out to OutputFile.

This change was made so conflicting files can be found before writing
the archive. E.g. take the following directory tree:

  -/bin
  -/bin/foo (file)
  -/bin/bar (file)

Then, make the following calls:

  AddFile(/bin, bin)
  AddFile(/bin/foo, bin/bar)

This conflict wouldn't be caught until writing out the archive, when it
could be caught much earlier.